// frontend/src/shared/lib/api/model/extensionFieldDefResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */
import type { ExtensionFieldDefResponseOptions } from './extensionFieldDefResponseOptions';

export interface ExtensionFieldDefResponse {
  id: string;
  kind: string;
  name: string;
  label: string;
  field_type: string;
  options: ExtensionFieldDefResponseOptions;
  position: number;
  show_in_table: boolean;
  version: number;
}