// frontend/src/shared/lib/api/model/extensionFieldType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */

export type ExtensionFieldType = typeof ExtensionFieldType[keyof typeof ExtensionFieldType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ExtensionFieldType = {
  string: 'string',
  text: 'text',
  number: 'number',
  integer: 'integer',
  boolean: 'boolean',
  date: 'date',
  enum: 'enum',
} as const;